Fixup matching, disallow repeating "types" (closes #23)
It is actually not a problem to have repeating instances of `links`,
`foreign_links` and `errors`, the items will just be appended.

To handle the zero-or-one rule for `types`, we can get away with just
one new matching rule, because there are three cases (no types, empty
types, and fully specified types).