Fix tests on stable/beta (closes #12)
Apparently, if the macro invocation is the last "expression" in the
function block, the parser expects an expression and will not let
the macro expand to anything else.  Appending a semicolon will
disabuse it of that notion.

The breakage only applies to error_chain! usage inside functions,
which is probably not relevant other than in these tests.

Seems to be fixed in nightly 1.12 anyway.