Modify the crate docs for bail!
The docs currently say that the bail! macro is like using try! or ?.
This isn't really correct. `try!` and `?` operate on a `Result` type,
returning the `Ok` value, if it is `Ok`, or returning early from the
function if it is an `Err`. `bail!` is equivalent to just returning
an `Err`, period. This change makes that a little more clear.