src: Remove executable bit from source files
I know this seems silly, but RPM builds of
https://github.com/coreos/coreos-installer/
are failing for me because RPM sticks these source files in
"debugsource" directories, and then a "linter" check notices
they appear to be executable but lack a shebang `#!` line.

There's no reason for source to be directly executable,
so remove that bit.